+++ saleor/graphql/shipping/types.py
@@ -4,6 +4,7 @@ from graphene import relay
 from ...core.permissions import ShippingPermissions
 from ...core.tracing import traced_resolver
 from ...core.weight import convert_weight_to_default_weight_unit
+from ...product import models as product_models
 from ...shipping import models
 from ...shipping.interface import ShippingMethodData
 from ..channel import ChannelQsContext
@@ -15,8 +16,12 @@ from ..channel.types import (
     ChannelContextTypeWithMetadata,
     ChannelContextTypeWithMetadataForObjectType,
 )
-from ..core.connection import CountableDjangoObjectType
-from ..core.fields import ChannelContextFilterConnectionField
+from ..core.connection import (
+    CountableConnection,
+    CountableDjangoObjectType,
+    create_connection_slice,
+)
+from ..core.fields import ConnectionField
 from ..core.types import CountryDisplay, Money, MoneyRange, Weight
 from ..decorators import permission_required
 from ..meta.types import ObjectWithMetadata
@@ -102,8 +107,8 @@ class ShippingMethodType(ChannelContextTypeWithMetadataForObjectType):
             "Postal code ranges rule of exclusion or inclusion of the shipping method."
         ),
     )
-    excluded_products = ChannelContextFilterConnectionField(
-        "saleor.graphql.product.types.products.Product",
+    excluded_products = ConnectionField(
+        "saleor.graphql.product.types.products.ProductCountableConnection",
         description="List of excluded products for the shipping method.",
     )
     minimum_order_weight = graphene.Field(
@@ -215,14 +220,18 @@ class ShippingMethodType(ChannelContextTypeWithMetadataForObjectType):
     @staticmethod
     @permission_required(ShippingPermissions.MANAGE_SHIPPING)
     def resolve_excluded_products(
-        root: ChannelContext[models.ShippingMethod], _info, **_kwargs
+        root: ChannelContext[models.ShippingMethod], info, **kwargs
     ):
-        if root.node.excluded_products is None:
-            return None
+        from ..product.types import ProductCountableConnection
 
-        return ChannelQsContext(
-            qs=root.node.excluded_products.all(), channel_slug=None  # type: ignore
-        )
+        if not root.node.excluded_products:
+            qs = product_models.Product.objects.none()
+        else:
+            qs = ChannelQsContext(
+                qs=root.node.excluded_products.all(), channel_slug=None  # type: ignore
+            )
+
+        return create_connection_slice(qs, info, kwargs, ProductCountableConnection)
 
 
 class ShippingZone(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
@@ -412,3 +421,8 @@ class ShippingMethod(graphene.ObjectType):
     @staticmethod
     def resolve_description(root: ShippingMethodData, *_args):
         return root.description
+
+
+class ShippingZoneCountableConnection(CountableConnection):
+    class Meta:
+        node = ShippingZone